re-organized sale folder into views and controllers
