+++ src/components/dashboard/DashBoard.js
@@ -3,7 +3,7 @@ import DataManager from "../../api/dataManager";
 import "../../styles/dashboard/dashboard.css"
 import SaleMetrics from "./saleMetrics"
 import Customers from "../customer/list.js";
-import Sales from "../sale/table.js";
+import Sales from "../sale/views/table.js";
 import Vehicles from "../vehicle/views/dashList.js";
 import Card from "@material-ui/core/Card";
 import CardContent from "@material-ui/core/CardContent";
