+++ src/components/sale/views/details.js
@@ -1,5 +1,5 @@
 import React from "react"
-import { modal } from "../../modules/modal/helpers"
+import { modal } from "../../../modules/modal/helpers"
 
 const Details = props => {
     return (
